Initial work on having seperate renderers so that a stereo/quad/5.1 can be split into seperate mono sources.
Plays as normal in Stereo mode but only in debug mode at the moment, need to work on stability as it doesn't seem too happy in release mode.

Need to add code to strip out the data - should see if this can be done via SGDMA to reduce CPU loading.


git-svn-id: svn+ssh://svn.code.sf.net/p/winampopenalout/code/trunk@65 b0487f5d-104f-0410-9801-bbc901c8d7b4